+++ R/features_method.R
@@ -29,7 +29,7 @@ setGeneric("features", function(x, y, ...) standardGeneric("features"))
 #' @references 
 #' Baker, Paul (2006): \emph{Using Corpora in Discourse Analysis}. London: continuum, p. 121-149 (ch. 6).
 #' 
-#' Manning, Christopher D.; Schütze, Hinrich (1999): \emph{Foundations of Statistical Natural Language
+#' Manning, Christopher D.; Schuetze, Hinrich (1999): \emph{Foundations of Statistical Natural Language
 #' Processing}. MIT Press: Cambridge, Mass., pp. 151-189 (ch. 5).
 #' @examples
 #' \dontrun{