+++ R/weigh_method.R
@@ -12,20 +12,20 @@ setGeneric("weigh", function(.Object, ...){standardGeneric("weigh")})
 #' @importFrom slam row_sums col_sums
 #' @importFrom tm nDocs
 #' @rdname weigh-method
-setMethod("weigh", "TermDocumentMatrix", function(.Object, method="tfidf"){
-  if (method=="tfidf"){
+setMethod("weigh", "TermDocumentMatrix", function(.Object, method = "tfidf"){
+  if (method == "tfidf"){
     .Object$v <- .Object$v/col_sums(.Object)[.Object$j] * log2(nDocs(.Object)/row_sums(.Object > 0))[.Object$i]  
     attr(.Object, "weighting") <- c(
       "term frequency - inverse document frequency (normalized)",
       "tf-idf"
     )
-  } else if (method=="rel"){
+  } else if (method == "rel"){
     .Object$v <- .Object$v/col_sums(.Object)[.Object$j]
     attr(.Object, "weighting") <- c(
       "term frequency (normalized)",
       "tf-normalized"
     )
-  } else if (method=="rank"){
+  } else if (method == "rank"){
     warning("not implemented")
   }
   return(.Object)
@@ -33,20 +33,20 @@ setMethod("weigh", "TermDocumentMatrix", function(.Object, method="tfidf"){
 
 
 #' @rdname weigh-method
-setMethod("weigh", "DocumentTermMatrix", function(.Object, method="tfidf"){
+setMethod("weigh", "DocumentTermMatrix", function(.Object, method = "tfidf"){
   if (method=="tfidf"){
     .Object$v <- .Object$v/row_sums(.Object)[.Object$i] * log2(nDocs(.Object)/col_sums(.Object > 0))[.Object$j]  
     attr(.Object, "weighting") <- c(
       "term frequency - inverse document frequency (normalized)",
       "tf-idf"
     )
-  } else if (method=="rel"){
+  } else if (method == "rel"){
     .Object$v <- .Object$v/row_sums(.Object)[.Object$i]
     attr(.Object, "weighting") <- c(
       "term frequency (normalized)",
       "tf-normalized"
     )
-  } else if (method=="rank"){
+  } else if (method == "rank"){
     warning("not implemented")
   }
   return(.Object)